requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-google-genai>=2.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
httpx>=0.27.0
supabase>=2.0.0
redis[hiredis]>=5.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
